update wpf client

+++ ChatProgramm-WebApp/chat/script.js
@@ -1,18 +1,23 @@
 //Autor: Stefan Rautner
-// URL zur MongoDB Datenbank definieren
-const urlToSpringBoot = 'http://localhost:8080/tinyWhatsApp';
 
 //Chat oder Gruppe hinzufügen
 async function createChat(event) {
     try {
         event.preventDefault();
-        await fetch(`${urlToSpringBoot}/newChat`, {
+        const response = await fetch(`${window.urlToSpringBoot}/newChat`, {
             method: 'POST',
             body: JSON.stringify({
                 'userID': window.userID,
                 'chatName': document.getElementById("new-chatname").value
             })
         });
+
+        if(!await response.text()) {
+            alert("Chat konnte nicht hinzugefügt werden");
+        } else {
+            window.location.href = '../home/home.html';
+            alert("Chat erfolgreich hinzugefügt");
+        }
     } catch (error) {
         console.error(error);
     }
@@ -22,12 +27,23 @@ async function createChat(event) {
 async function deleteChat(event) {
     try {
         event.preventDefault();
-        await fetch(`${urlToSpringBoot}/deleteChat`, {
-            method: 'DELETE',
-            body: JSON.stringify({
-                'chatID': window.chatID
-            })
-        });
+        if(chatID !== null && chatID !== "") {
+            const response = await fetch(`${window.urlToSpringBoot}/deleteChat`, {
+                method: 'DELETE',
+                body: JSON.stringify({
+                    'chatID': window.chatID
+                })
+            });
+
+            if(!await response.text()) {
+                alert("Chat konnte nicht hinzugefügt werden");
+            } else {
+                window.location.href = '../home/home.html';
+                alert("Chat erfolgreich gelöscht");
+            }
+        } else {
+            alert("Bitte wählen Sie einen Chat aus");
+        }
     } catch (error) {
         console.error(error);
     }
@@ -36,13 +52,24 @@ async function deleteChat(event) {
 async function updateChatNames(event) {
     try {
         event.preventDefault();
-        await fetch(`${urlToSpringBoot}/updateChatName`, {
-            method: 'PUT',
-            body: JSON.stringify({
-                'chatID': window.chatID,
-                'chatName': document.getElementById("new-chatname").value
-            })
-        });
+        if(chatID !== null && chatID !== "") {
+            const response = await fetch(`${window.urlToSpringBoot}/updateChatName`, {
+                method: 'PUT',
+                body: JSON.stringify({
+                    'chatID': window.chatID,
+                    'chatName': document.getElementById("new-chatname").value
+                })
+            });
+
+            if(!await response.text()) {
+                alert("Chatname konnte nicht aktualisiert werden");
+            } else {
+                window.location.href = '../home/home.html';
+                alert("Chatname wurde erfolgreich aktualisiert");
+            }
+        } else {
+            alert("Bitte wählen Sie einen Chat aus");
+        }
     } catch (error) {
         console.error(error);
     }
@@ -51,13 +78,23 @@ async function updateChatNames(event) {
 async function addUserToChat(event) {
     try {
         event.preventDefault();
-        await fetch(`${urlToSpringBoot}/addUserToChat`, {
-            method: 'POST',
-            body: JSON.stringify({
-                'username': document.getElementById("add-user").value,
-                'chatID': window.chatID
-            })
-        });
+        if(chatID !== null && chatID !== "") {
+            const response = await fetch(`${window.urlToSpringBoot}/addUserToChat`, {
+                method: 'POST',
+                body: JSON.stringify({
+                    'username': document.getElementById("add-user").value,
+                    'chatID': window.chatID
+                })
+            });
+
+            if(!await response.text()) {
+                alert("Benutzer konnte nicht zum Chat hinzugefügt werden");
+            } else {
+                alert("Benutzer wurde erfolgreich in den Chat hinzugefügt");
+            }
+        } else {
+            alert("Bitte wählen Sie einen Chat aus");
+        }
     } catch (error) {
         console.error(error);
     }
@@ -66,13 +103,23 @@ async function addUserToChat(event) {
 async function removeUserFromChat(event) {
     try {
         event.preventDefault();
-        await fetch(`${urlToSpringBoot}/removeUserFromChat`, {
-            method: 'POST',
-            body: JSON.stringify({
-                'username': document.getElementById("add-user").value,
-                'chatID': window.chatID
-            })
-        });
+        if(chatID !== null && chatID !==  "") {
+            const response = await fetch(`${window.urlToSpringBoot}/removeUserFromChat`, {
+                method: 'POST',
+                body: JSON.stringify({
+                    'username': document.getElementById("add-user").value,
+                    'chatID': window.chatID
+                })
+            });
+
+            if(!await response.text()) {
+                alert("Benutzer konnte nicht aus dem Chat entfernt werden");
+            } else {
+                alert("Benutzer wurde erfolgreich aus dem Chat entfernt");
+            }
+        } else {
+            alert("Bitte wählen Sie einen Chat aus");
+        }
     } catch (error) {
         console.error(error);
     }
